+++ apps/nextjs/src/components/Feedback/index.tsx
@@ -44,7 +44,7 @@ const FeedBack = ({
       {/* Close Button */}
 
       <form
-        onSubmit={async (e) => {
+        onSubmit={(e) => {
           e.preventDefault();
         }}
         className="flex w-full flex-col gap-14"
@@ -93,11 +93,11 @@ const FeedBack = ({
                       }}
                     >
                       <span
-                        className={`rounded-sm border-2  p-8 px-9 text-lg sm:px-15 ${
+                        className={`rounded-sm border-2 p-8 px-9 text-lg sm:px-15 ${
                           usersResponse[surveyResponseKey] ===
                           feedback.number.toString()
                             ? "border-black bg-black text-white"
-                            : " border-oakGrey3 bg-white text-black"
+                            : "border-oakGrey3 bg-white text-black"
                         }`}
                       >
                         {feedback.number}