+++ packages/rag/jest.config.mjs
@@ -1,0 +1,22 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+const config = {
+  transform: {
+    "^.+\\.tsx?$": [
+      "ts-jest",
+      {
+        tsconfig: "tsconfig.test.json",
+        useESM: true,
+        isolatedModules: true,
+      },
+    ],
+  },
+  preset: "ts-jest/presets/default-esm",
+  extensionsToTreatAsEsm: [".ts"],
+  testEnvironment: "setup-polly-jest/jest-environment-node",
+  testMatch: ["**/*.test.ts"],
+  moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+  rootDir: ".",
+  resetMocks: true,
+};
+
+export default config;